feat(UnlockTier): unlock tier if locked at upgrade
After an upgrade attempt has been made, it recognizes if the TowerXP
menu has been opened due to the tier not being unocked, unlocks it, and
reattempts to upgrade the tier.

Disabled by default as it should only be used when you don't have
certain tiers unlocked that get used within the strats.

To use: set enableUnlockTier=true in config.ini