refactor(maps): monkey_meadow works w/low towerxp
With the addition of UnlockTier(), refactored monkey_meadow hard.ahk to
work even if the user has no tower xp for any towers or any monkey
knowledge.

Additionally, it gives enough xp to the towers used in GeneralHard(),
allowing a user run any map that utilizes the strat no matter their
progress in the game.