fix(ok button): more strict and versatile
The ok button was triggering sometimes when it shouldn't because it was
all white. Made a new ok.png using my masking technique to make a more
strict search that remains versatile. Tested and confirmed to work on
most if not every place you come across an ok button. Specifically:
    - event ended message on home screen
    - br/ct info page
    - first ever login "welcome home" message
    - new update message
    - in game interrupts
    - internet unavailable message on home/team page